% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{add_fac_unavail}
\alias{add_fac_unavail}
\title{Add vaculty unavailability to schedule}
\usage{
add_fac_unavail(f_unavail, f_schedule)
}
\arguments{
\item{f_unavail}{dataframe of faculty unavailability (columns: Faculty, Slot)}

\item{f_schedule}{faculty schedule (generated with make_faculty_schedule)}
}
\value{
faculty schedule with unavailability documented
}
\description{
Add vaculty unavailability to schedule
}
