% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{add_min_fac_meetings}
\alias{add_min_fac_meetings}
\title{Ad minimum nuber of faculty meetings}
\usage{
add_min_fac_meetings(
  s_schedule,
  ranked_faculty,
  min_fac_mtg = nrow(s_schedule)/2,
  f_unavail = NULL
)
}
\arguments{
\item{s_schedule}{student schedule}

\item{ranked_faculty}{data frame of ranked faculty}

\item{min_fac_mtg}{minimum number of faculty meetings}
}
\value{
updated student schedule
}
\description{
Ad minimum nuber of faculty meetings
}
