% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{already_met}
\alias{already_met}
\title{Remove faculty the student already met with}
\usage{
already_met(am, rf, faculty)
}
\arguments{
\item{am}{dataframe with student names as the first column and faculty names as the second column. Student names can be repeated.}

\item{rf}{dataframe with student names as column names}

\item{faculty}{dataframe where each row is a faculty member and each column is a research interest. First column is faculty names.}
}
\value{

}
\description{
Remove faculty the student already met with
}
