% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fac_avail.R
\name{check_fac_avail}
\alias{check_fac_avail}
\title{Check faculty availability}
\usage{
check_fac_avail(f_unavail, f_name, slot)
}
\arguments{
\item{f_unavail}{dataframe of faculty unavailability (columns: Faculty, Slots)}

\item{f_name}{name of faculty to check}

\item{slot}{time slot to check}
}
\value{

}
\description{
Check faculty availability
}
