% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_scores.R
\name{get_match_scores}
\alias{get_match_scores}
\title{Get match scores}
\usage{
get_match_scores(students, faculty)
}
\arguments{
\item{students}{dataframe where each row is a student and each column is a research interest. First column is student names.}

\item{faculty}{dataframe where each row is a faculty member and each column is a research interest. First column is faculty names.}
}
\value{
dataframe where rows are students and columns are faculty and elements are a score of how well the student and faculty research interests align. Higher scores indicate more mutual research interests.
}
\description{
Get match scores
}
