% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{make_faculty_schedule}
\alias{make_faculty_schedule}
\title{Make faculty schedule}
\usage{
make_faculty_schedule(student_schedule, f_unavail)
}
\arguments{
\item{student_schedule}{student schedule (generated with make_student_schedule)}

\item{f_unavail}{dataframe of faculty unavailability (name, time slot)}
}
\value{
faculty schedule
}
\description{
Make faculty schedule
}
