% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_student_sched.R
\name{make_student_schedule}
\alias{make_student_schedule}
\title{Make student schedule}
\usage{
make_student_schedule(ranked_faculty, slots = 12, f_unavail = NULL)
}
\arguments{
\item{ranked_faculty}{dataframe returned from rank_faculty}

\item{slots}{number of slots in schedule}

\item{f_unavail}{data frame of whether faculty are available to meet (columns: Faculty, Slot)}
}
\value{
student schedule
}
\description{
Make student schedule
}
