% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchathon.R
\name{matchathon}
\alias{matchathon}
\title{Matchathon master function}
\usage{
matchathon(
  faculty_df,
  student_df,
  meeting_slots = 12,
  min_fslots = NULL,
  f_unavail_df = NULL,
  already_met_df = NULL
)
}
\arguments{
\item{faculty_df}{csv of faculty interests}

\item{student_df}{csv of student interests}

\item{meeting_slots}{number of meeting slots}

\item{min_fslots}{minimum number of faculty meeting slots not empty}

\item{f_unavail_df}{dataframe of faculty unavailability (name, time slot)}

\item{already_met_df}{dataframe of faculty students have already met with (student name, faculty name)}
}
\value{
faculty and student schedules and ranked faculty lists for students
}
\description{
Matchathon master function
}
