% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_faculty.R
\name{rank_faculty}
\alias{rank_faculty}
\title{Rank faculty for each student}
\usage{
rank_faculty(match_score)
}
\arguments{
\item{match_score}{dataframe returned by get_match_scores}
}
\value{
For each student (column), faculty with mutual interests ordered by match score.
}
\description{
Rank faculty for each student
}
