% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_data.R
\name{read_in_data}
\alias{read_in_data}
\title{Read in csvs}
\usage{
read_in_data(
  faculty_csv,
  student_csv,
  f_unavail_csv = NULL,
  already_met_csv = NULL
)
}
\arguments{
\item{faculty_csv}{csv of faculty interests}

\item{student_csv}{csv of student interests}

\item{f_unavail_csv}{csv of faculty unavailability}

\item{already_met_csv}{csv of faculty students have already met with}
}
\value{
list with dataframes for each csv
}
\description{
Read in csvs
}
