% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{matchathon}
\alias{matchathon}
\title{Matchathon master function}
\usage{
matchathon(
  faculty_csv,
  students_csv,
  meeting_slots = 12,
  min_fslots = NULL,
  f_unavail = NULL
)
}
\arguments{
\item{faculty_csv}{csv of faculty interests}

\item{students_csv}{csv of student interests}
}
\value{
faculty and student schedules and ranked faculty lists for students
}
\description{
Matchathon master function
}
